+++ src/features/Conversation/Error/index.tsx
@@ -1,78 +1,83 @@
-import { PluginErrorType } from '@lobehub/chat-plugin-sdk';
+import { IPluginErrorType, PluginErrorType } from '@lobehub/chat-plugin-sdk';
+import type { AlertProps } from '@lobehub/ui';
+import { memo } from 'react';
 
-import { ChatErrorType } from '@/types/fetch';
+import { AgentRuntimeErrorType, ILobeAgentRuntimeErrorType } from '@/libs/agent-runtime';
+import { ChatErrorType, ErrorType } from '@/types/fetch';
+import { ChatMessage, ChatMessageError } from '@/types/message';
 
-import { RenderErrorMessage } from '../types';
-import InvalidAccess from './InvalidAccess';
-import OpenAPIKey from './OpenAPIKey';
+import ErrorJsonViewer from './ErrorJsonViewer';
+import InvalidAPIKey from './InvalidAPIKey';
+import InvalidAccessCode from './InvalidAccessCode';
 import OpenAiBizError from './OpenAiBizError';
-import PluginError from './Plugin/PluginError';
-import PluginSettings from './Plugin/PluginSettings';
+import PluginSettings from './PluginSettings';
 
-export const renderErrorMessages: Record<string, RenderErrorMessage> = {
-  [PluginErrorType.PluginMarketIndexNotFound]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginMarketIndexInvalid]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginMetaInvalid]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginMetaNotFound]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginManifestInvalid]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginManifestNotFound]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginApiNotFound]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginApiParamsError]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginServerError]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginGatewayError]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginOpenApiInitError]: {
-    Render: PluginError,
-  },
-  [PluginErrorType.PluginSettingsInvalid]: {
-    Render: PluginSettings,
-    config: {
+// Config for the errorMessage display
+export const getErrorAlertConfig = (
+  errorType?: IPluginErrorType | ILobeAgentRuntimeErrorType | ErrorType,
+): AlertProps | undefined => {
+  // OpenAIBizError / ZhipuBizError / GoogleBizError / ...
+  if (typeof errorType === 'string' && errorType.includes('Biz'))
+    return {
       extraDefaultExpand: true,
       extraIsolate: true,
       type: 'warning',
-    },
-  },
-  [ChatErrorType.InvalidAccessCode]: {
-    Render: InvalidAccess,
-    config: {
-      extraDefaultExpand: true,
-      extraIsolate: true,
-      type: 'warning',
-    },
-  },
-  [ChatErrorType.NoAPIKey]: {
-    Render: OpenAPIKey,
-    config: {
-      extraDefaultExpand: true,
-      extraIsolate: true,
-      type: 'warning',
-    },
-  },
-  [ChatErrorType.OpenAIBizError]: {
-    Render: OpenAiBizError,
-    config: {
-      extraDefaultExpand: true,
-      extraIsolate: true,
-      type: 'warning',
-    },
-  },
+    };
+
+  switch (errorType) {
+    case AgentRuntimeErrorType.LocationNotSupportError: {
+      return {
+        type: 'warning',
+      };
+    }
+
+    case PluginErrorType.PluginSettingsInvalid:
+    case ChatErrorType.InvalidAccessCode:
+    case AgentRuntimeErrorType.NoOpenAIAPIKey:
+    case AgentRuntimeErrorType.InvalidBedrockCredentials:
+    case AgentRuntimeErrorType.InvalidGoogleAPIKey:
+    case AgentRuntimeErrorType.InvalidZhipuAPIKey: {
+      return {
+        extraDefaultExpand: true,
+        extraIsolate: true,
+        type: 'warning',
+      };
+    }
+
+    default: {
+      return undefined;
+    }
+  }
 };
+
+const ErrorMessageExtra = memo<{ data: ChatMessage }>(({ data }) => {
+  const error = data.error as ChatMessageError;
+  if (!error?.type) return;
+
+  switch (error.type) {
+    case PluginErrorType.PluginSettingsInvalid: {
+      return <PluginSettings id={data.id} plugin={data.plugin} />;
+    }
+
+    case AgentRuntimeErrorType.OpenAIBizError: {
+      return <OpenAiBizError {...data} />;
+    }
+
+    case ChatErrorType.InvalidAccessCode: {
+      return <InvalidAccessCode id={data.id} provider={data.error?.body?.provider} />;
+    }
+
+    case AgentRuntimeErrorType.InvalidBedrockCredentials:
+    case AgentRuntimeErrorType.InvalidZhipuAPIKey:
+    case AgentRuntimeErrorType.InvalidGoogleAPIKey:
+    case AgentRuntimeErrorType.NoOpenAIAPIKey: {
+      return <InvalidAPIKey id={data.id} provider={data.error?.body?.provider} />;
+    }
+
+    default: {
+      return <ErrorJsonViewer error={data.error} id={data.id} />;
+    }
+  }
+});
+
+export default ErrorMessageExtra;