+++ src/libs/agent-runtime/types/type.ts
@@ -1,0 +1,34 @@
+import OpenAI from 'openai';
+
+import { ILobeAgentRuntimeErrorType } from '../error';
+import { ChatStreamPayload } from './chat';
+
+export interface AgentInitErrorPayload {
+  error: object;
+  type: string | number;
+}
+
+export interface ChatCompletionErrorPayload {
+  [key: string]: any;
+  endpoint?: string;
+  error: object;
+  errorType: ILobeAgentRuntimeErrorType;
+  provider: ModelProvider;
+}
+
+export interface CreateChatCompletionOptions {
+  chatModel: OpenAI;
+  payload: ChatStreamPayload;
+}
+
+export enum ModelProvider {
+  Anthropic = 'anthropic',
+  Azure = 'azure',
+  Bedrock = 'bedrock',
+  ChatGLM = 'chatglm',
+  Google = 'google',
+  Mistral = 'mistral',
+  OpenAI = 'openai',
+  Tongyi = 'tongyi',
+  ZhiPu = 'zhipu',
+}